1，功法加技能说明 2，战斗胜利奖励银两会加入仓库BUG

+++ Classes/GlobalInstance.h
@@ -73,7 +73,7 @@ typedef struct
 	std::vector<int> vec_skillbns;//技能加成
 	int skillrnd;
 	float skilleff1;
-	float skilleff2;
+	int skilleff2;
 	int skill;
 	int qu;
 	std::vector<int> vec_hp;
